Rename plugins.lua to load-plugins.lua
